requirements.txt
python-telegram-bot==13.1.0
requests~=2.25.1
SpeechRecognition==3.8.1
curses-menu==0.5.0
beautifulsoup4==4.9.3
ffmpeg==4.2.2